modified the useResourcesApi hook
